第四天代码进度

+++ yy-club-subject/yy-club-infra/src/main/java/com/jingdianyy/subject/infra/basic/entity/SubjectBrief.java
@@ -20,7 +20,7 @@ public class SubjectBrief implements Serializable {
     /**
      * 题目id
      */
-    private Integer subjectId;
+    private Long subjectId;
     /**
      * 题目答案
      */
